// gen/esi.ts
export namespace esi {
// ...
    export namespace character {
// ...
        export namespace mail {
// ...
            export interface List {
                /**
                 * Mailing list ID.
                 */
                mailing_list_id: number;
                name: string;
            }
            /**
             * This is the response type for the route, [`GET /v1/characters/{character_id}/mail/{mail_id}/`](https://esi.evetech.net//#!/Mail/get_characters_character_id_mail_mail_id).
             */
            export interface Mail {
                body?: string;
                /**
                 * From whom the mail was sent.
                 */
                from?: number;
                /**
                 * Labels attached to the mail.
                 */
                labels?: number[];
                /**
                 * Whether the mail is flagged as read.
                 */
                read?: boolean;
                /**
                 * Recipients of the mail.
                 */
                recipients?: esi.character.mail.Recipient[];
                subject?: string;
                /**
                 * When the mail was sent.
                 */
                timestamp?: string;
            }
            export interface MailHeader {
                /**
                 * From whom the mail was sent.
                 */
                from?: number;
                is_read?: boolean;
                labels?: number[];
                mail_id?: number;
                /**
                 * Recipients of the mail.
                 */
                recipients?: esi.character.mail.Recipient[];
                subject?: string;
                /**
                 * When the mail was sent.
                 */
                timestamp?: string;
            }
// ...
            export interface Recipient {
                recipient_id: number;
                recipient_type: esi.character.mail.RecipientType;
            }
            export const enum RecipientType {
                ALLIANCE = "alliance",
                CHARACTER = "character",
                CORPORATION = "corporation",
                MAILING_LIST = "mailing_list"
            }
        }
// ...
    }
// ...
}
// ...
export interface Responses {
// ...
    delete_characters_character_id_mail_labels_label_id: undefined;
    /**
     * The type of this member is the response type of for the route: [`DELETE /v1/characters/{character_id}/mail/{mail_id}/`](https://esi.evetech.net//#!/Mail/delete_characters_character_id_mail_mail_id).
     */
    delete_characters_character_id_mail_mail_id: undefined;
    /**
     * The type of this member is the response type of for the route: [`GET /v1/characters/{character_id}/mail/`](https://esi.evetech.net//#!/Mail/get_characters_character_id_mail).
     */
    get_characters_character_id_mail: esi.character.mail.MailHeader[];
// ...
    get_characters_character_id_mail_lists: esi.character.mail.List[];
    /**
     * The type of this member is the response type of for the route: [`GET /v1/characters/{character_id}/mail/{mail_id}/`](https://esi.evetech.net//#!/Mail/get_characters_character_id_mail_mail_id).
     */
    get_characters_character_id_mail_mail_id: esi.character.mail.Mail;
    /**
     * The type of this member is the response type of for the route: [`POST /v1/characters/{character_id}/mail/`](https://esi.evetech.net//#!/Mail/post_characters_character_id_mail).
     */
    post_characters_character_id_mail: number;
    /**
     * The type of this member is the response type of for the route: [`POST /v2/characters/{character_id}/mail/labels/`](https://esi.evetech.net//#!/Mail/post_characters_character_id_mail_labels).
     */
    post_characters_character_id_mail_labels: number;
    /**
     * The type of this member is the response type of for the route: [`PUT /v1/characters/{character_id}/mail/{mail_id}/`](https://esi.evetech.net//#!/Mail/put_characters_character_id_mail_mail_id).
     */
    put_characters_character_id_mail_mail_id: undefined;
// ...
}